requirements.txt
Flask>=3.0
SQLAlchemy>=2.0
PyMySQL>=1.1
pandas>=2.1
python-dotenv>=1.0